+++ packages/gladia/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/gladia
 
+## 2.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [3b1d015]
+  - @ai-sdk/provider-utils@4.0.0-beta.14
+
+## 2.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [d116b4b]
+  - @ai-sdk/provider-utils@4.0.0-beta.13
+
 ## 2.0.0-beta.12
 
 ### Patch Changes