+++ packages/google/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/google
 
+## 3.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [3b1d015]
+  - @ai-sdk/provider-utils@4.0.0-beta.14
+
+## 3.0.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies [d116b4b]
+  - @ai-sdk/provider-utils@4.0.0-beta.13
+
 ## 3.0.0-beta.17
 
 ### Patch Changes