+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/groq
 
+## 3.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [3b1d015]
+  - @ai-sdk/provider-utils@4.0.0-beta.14
+
+## 3.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [d116b4b]
+  - @ai-sdk/provider-utils@4.0.0-beta.13
+
 ## 3.0.0-beta.12
 
 ### Patch Changes