+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,25 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.40
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [f6f0c5a]
+  - ai@6.0.0-beta.39
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.38
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@6.0.0-beta.37
 
 ## 0.0.1-beta.0